Documents/mlh-project/flask-server/requirements.txt
PyPDF2
dotenv
google.generativeai